dom/media/test/test_play_promise_1.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: promise-based play() method</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">

<script>
// Name: playBeforeCanPlay
// Case: invoke play() on an element that doesn't have enough data
// Expected result: resolve the promise

let manager = new MediaTestManager;

function initTest(test, token) {
  manager.started(token);

  let element = document.createElement(getMajorMimeType(test.type));
  element.src = test.name;
  ok(element.readyState == HTMLMediaElement.HAVE_NOTHING);

  element.play().then(
    (result) => {
      if (result == undefined) {
        ok(true, `${token} is resolved with ${result}.`);
      } else {
        ok(false, `${token} is resolved with ${result}.`);
      }
    },
    (error) => {
      ok(false, `${token} is rejected with ${error.name}.`);
    }
  ).then( () => { manager.finished(token); } );
}

manager.runTests(gSmallTests, initTest);

</script>
